Added List command, and renamed example.js
***example.js was renamed to main.js
Mostly added the list command, the formatting isn't good, and it doesn't return all the required fields yet. Right now the user it gets issues from is hard-coded, you can change it by modifying the "user" var. This commit has a change to the dockerfile to allow the bot to use the node-fetch library, so when you run this version for the first time make sure to rebuild it (docker-compose build). The function that sorts through the information returned by Jira, is in the file getList.js in the scripts directory. It's exported as a module so that it can be used by the main file but still be in a separate file. Right now the way it's set up, each (Jira related) function will need to be in its own file, which would mean something like having one file for each command, but I'm working on merging all the Jira related stuff into one file, so it would be one file for Jira stuff (including text processing and whatnot), and one file for bot stuff.